+++ packages/jaspr/test/browser/basic/basic_app.dart
@@ -38,10 +38,6 @@ class Button extends StatelessComponent {
 
   @override
   Iterable<Component> build(BuildContext context) sync* {
-    yield DomComponent(
-      tag: 'button',
-      child: Text(label),
-      events: {'click': (e) => onPressed()},
-    );
+    yield button(events: {'click': (e) => onPressed()}, [text(label)]);
   }
 }