+++ packages/jaspr/test/server/without_jaspr_testing_library/without_jaspr_testing_library_app.dart
@@ -3,8 +3,8 @@ import 'package:jaspr/jaspr.dart';
 class App extends StatelessComponent {
   @override
   Iterable<Component> build(BuildContext context) sync* {
-    yield DomComponent(tag: 'div', children: [
-      Text('Hello'),
+    yield div([
+      text('Hello'),
     ]);
   }
 }